backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.1
django-cors-headers==4.4.0
Pillow==10.4.0
celery==5.4.0
redis==5.0.8
reportlab==4.2.2
python-decouple==3.8
qrcode==7.4.2
django-filter==24.3
scikit-learn==1.5.1
numpy==2.0.1
